+++ src/music/meta/MusicMeta.tsx
@@ -8,15 +8,35 @@ export default function MusicMeta(props: {
   keySignature?: IKeySignature
   timeSignature?: ITimeSignature
   composerHtml?: string
+  onKeySignatureChange?: (value?: IKeySignature) => void
+  onTimeSignatureChange?: (value?: ITimeSignature) => void
   onComposerChange?: (value: string) => void
 }) {
-  const { keySignature, timeSignature, composerHtml, onComposerChange } = props
+  const {
+    keySignature,
+    timeSignature,
+    composerHtml,
+    onKeySignatureChange,
+    onTimeSignatureChange,
+    onComposerChange,
+  } = props
 
   return (
     <div className='music-meta'>
-      <div>
-        {keySignature && <KeySignature value={keySignature}></KeySignature>}
-        {timeSignature && <TimeSignature value={timeSignature}></TimeSignature>}
+      <div className='music-meta-left'>
+        {keySignature && (
+          <KeySignature
+            value={keySignature}
+            onChange={onKeySignatureChange}
+          ></KeySignature>
+        )}
+        <div style={{ width: '12px' }}></div>
+        {timeSignature && (
+          <TimeSignature
+            value={timeSignature}
+            onChange={onTimeSignatureChange}
+          ></TimeSignature>
+        )}
       </div>
       <div>
         {composerHtml && (